refactor: 무한 스크롤 로직 공통화 - useInfiniteScroll 훅 생성 및 적용
