+++ app/list/page.tsx
@@ -1,12 +1,9 @@
 'use client';
 
-/* eslint-disable compat/compat */
-
-import { useEffect, useRef } from 'react';
-
 import classNames from 'classnames/bind';
 
 import Card from '@/app/list/_components/Card/Card';
+import { useInfiniteScroll } from '@/src/shared/hooks/useInfiniteScroll';
 import { UrlString } from '@apis/types/Recipient';
 import { ColorchipColors } from '@constants/COLORS';
 import { useRecipientsListInfinite } from '@queries/useRecipientQueries';
@@ -35,10 +32,11 @@ export default function List() {
     isFetchingNextPage: isFetchingNextRecent,
   } = useRecipientsListInfinite({ limit: 8, sort: 'recent' });
 
-  const popularObserverRef = useRef<IntersectionObserver | null>(null);
-  const recentObserverRef = useRef<IntersectionObserver | null>(null);
-  const loadMorePopularRef = useRef<HTMLDivElement>(null);
-  const loadMoreRecentRef = useRef<HTMLDivElement>(null);
+  const loadMorePopularRef = useInfiniteScroll(fetchNextPopular, hasNextPopular, isFetchingNextPopular, {
+    threshold: 1,
+  });
+
+  const loadMoreRecentRef = useInfiniteScroll(fetchNextRecent, hasNextRecent, isFetchingNextRecent, { threshold: 1 });
 
   // 인기 캐러셀이 끝에 도달했을 때 호출되는 핸들러
   const handleReachEndPopular = () => {
@@ -53,56 +51,6 @@ export default function List() {
       fetchNextRecent();
     }
   };
-
-  useEffect(() => {
-    if (popularObserverRef.current) {
-      popularObserverRef.current.disconnect();
-    }
-
-    popularObserverRef.current = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPopular && !isFetchingNextPopular) {
-          fetchNextPopular();
-        }
-      },
-      { threshold: 1 },
-    );
-
-    if (loadMorePopularRef.current) {
-      popularObserverRef.current.observe(loadMorePopularRef.current);
-    }
-
-    return () => {
-      if (popularObserverRef.current) {
-        popularObserverRef.current.disconnect();
-      }
-    };
-  }, [fetchNextPopular, hasNextPopular, isFetchingNextPopular]);
-
-  useEffect(() => {
-    if (recentObserverRef.current) {
-      recentObserverRef.current.disconnect();
-    }
-
-    recentObserverRef.current = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextRecent && !isFetchingNextRecent) {
-          fetchNextRecent();
-        }
-      },
-      { threshold: 1 },
-    );
-
-    if (loadMoreRecentRef.current) {
-      recentObserverRef.current.observe(loadMoreRecentRef.current);
-    }
-
-    return () => {
-      if (recentObserverRef.current) {
-        recentObserverRef.current.disconnect();
-      }
-    };
-  }, [fetchNextRecent, hasNextRecent, isFetchingNextRecent]);
 
   // 인기 롤링페이퍼(popular) 데이터
   const popularRecipients = popularData?.pages.flatMap((page) => page.results) || [];
